Reload tenant configuration on change merged
Whenever a change that updates the zuul config merges, Zuul should
immediately update its configuration.  This should be generally
a no-op for changes to project repos in a dependent pipeline,
however, it may affect indepedent pipelines of project repos or
anything related to config repos (for which the config is not
updated dynamically).

To do this efficiently, try to use as much cached data as possible,
but still perform the re-enqueue operations which are normal for
a reconfiguration to ensure correctness.  Limit the reconfiguration
to just the affected tenant.  NOTE: this may be an error -- tenants
may share common config repos, and therefore, if a change lands to
such a repo, more than one tenant may need to be updated.

Change-Id: Ib35b0b663423b1a27f9e3bcdec7480345e3bdff1