Bump post playbook timeout to 30mins
Currently our post playbook timeout value is hardcoded to 10mins, for
the majority of our jobs this is okay. However, when projects need to
transfer a lot of data (kolla 2.6gb tarballs) zuul will abort the post
playbook.

For zuulv3, we should properly expose this value to be configured per
job, but today just bump our timeout to 30mins.

Change-Id: I12dcbfe60bb1d59c3af8a13f49f04e3b68ff7197
Signed-off-by: Paul Belanger <pabelanger@redhat.com>