Set display_args_to_stdout for secure jobs
Ansible has a config flag that is accessible from callback plugins that
is an indicator for whether or not to display args in the log output.

It should be noted that this can wind up being a general flag in our
callback plugin that will let us know if the job being run is secure or
insecure.

Change-Id: Ie0b45ca533e71610cc18950edd735dc3258bd604