Simplify the log url
Drop some of the hard-coded path expectations.

It turns out we can do everything we were doing in the hardcoded
method using string format operations.

This reduces some of the variables that are available for URL
formatting.  Useful variables are made available to the formatters
in a safe manner (so that users can't access job.auth.secrets...).

Co-Authored-By: James E. Blair <jeblair@redhat.com>
Change-Id: Ide646d918f4b8e54658809feaa3a8f5ccf228ebc