Report dynamic layout config errors
If Zuul encounters a syntax error when loading an in-repo layout,
report that as a failure on the change.  Try to provide as much
readable information as possible to allow the user to diagnose the
error.

All top-level configuration objects are updated to support a hidden
source context variable to aid in creating a useful error message.

Item.areAllJobsComplete is updated so that it returns true in the
case of a merge failure or config error.  In Zuulv2, we knew what
jobs we would run even in those cases, but in Zuulv3, we don't, so
the item does not have a job tree, so in those cases, the item does
not report that all jobs are complete.  Returining True in this case
allows the NNFI algorithm to continue and expel the item from the
queue, avoiding an inifinite loop.

The merge failure accessor method is simplified.

Change-Id: I9cf19b87c6af5926b5e8bb403b81ce0470e3592d