Remove layoutvalidator
This is no longer used.  Instead, the configloader is reponsible
for validating each configuration stanza as it is parsed.  The
error reporting happens as changes are proposed to in-repo
configuration files.  Most of the good paths are tested by the
tests which utilize those features.  There are a few tests which
deliberately test syntax errors, though they are not as
comprehensive.  Because a nearly complete running Zuul system is
required in order to parse a configuration file now, it is not
practical to test configuration syntax errors in the way that was
previously done in the layoutvalidator; therefore the tests
are removed without a replacement.

Change-Id: I774dccbc2460099fc24815ed96f0de7532157d31