Fix dynamic reconfiguration
The dynamic reconfiguration would omit portions of the configuration
which originated from config-repos.  Specifically, pipeline and
semaphore configuration would be copied over because they can not
be changed in dynamic reconfiguration, and are handles as special
cases.  However, all job and project configuration which appeared
in a config-repo would be omitted.  We didn't notice this because
most of our in-repo reconfiguration did not rely on jobs defined
in a config-repo (only project-repos).  The following change does
rely on this and exposed the issue.

Change-Id: Ibaf2d62a6f40bcb1ea96d260d96d6518ee511b85