Extend test timeout to 120s
This may help tests pass when running under 4 vcpus.  Also:

A following change adds more git operations to most of the tests.
This is because each git repo must be listed in the tenant config
file, and many tests share a tenant config with quite a number
of repos which they don't use.  We should improve the tests so that
they don't operate on repos that they don't need, but that will
take some work.

Change-Id: I91a327968359b3650fbe0477e99dc274a1fb6266