Configure zk fixture cleanup early
Generally you want to put the cleanup in place before the resource is
created. This way if you have failure in setUp then you don't leak the
resources as they should be cleaned up after early failures. Move the zk
fixture's chroot cleanup to before we create the chroot. This way if
anything fails in setting that up we should get cleaned up properly.

Also explicitly close the tmp kazoo client after stopping it as kazoo
docs say not doing this can result in leaks.

Change-Id: I6d94909b5b4836aa09528e1de744c940c2a9a12f