Don't allow auth definition in job to be a list
In experimenting with secrets and reading the configloader it appears
like the auth option is supposed to be a list. However when auth is
provided as a list later handling fails and an 'Unknown configuration
error' is thrown.

I think the default here should be that auth is a dictionary if it is
provided.

Change-Id: I573ad0d819905c0351e7d50bcc3c7b9c4d03e2d1
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>