Add simple_layout test decorator
And add it to test_no_job_project.

So that we can reduce the number of boilerplate git repos initialized
for every (or most) tests, make it simple to specify a single simple
layout file (like we used to do in zuul v2).  The more complicated
scenarios can rely on the full copy-git-repos-from-test-fixtures
behavior, but simple tests can have that automatically generated from
a simple zuul.yaml file.

Change-Id: Ibeec4c526a1097823589f2c38a50e40dc346e0e5