Remove unused repos from repo init
node-project and conflict-project weren't used at all.
layered-project and templated-project are now in a
different config tree.
project4, 5, and 6 are only used by the cloner tests which have
not been ported to v3 (and may not be for a while since the
current incarnation of the cloner, as such, isn't needed).

Change-Id: If522ad9245929d96b6e3f6a29921562fa830edea