Also send stdout back in the Result object
It's a legitimate usecase in ansible to register the results of a shell
command and then test its value and make decisions. Not sending stdout
back in the result object would make playbooks that do that break.

There is a risk that we die under the load of really long log files
though - so let's keep track of that.

Change-Id: I2c2fe558d6ec93cef7bb4b1c5abd983488edd745